+++ R/data.R
@@ -53,3 +53,16 @@
 #' @family datasets
 #' @references https://github.com/anjastein/SMC-Mallows/tree/main/data
 "sample_dataset"
+
+#' Updated partial rankings of potatoes
+#'
+#' Simulated dataset based on \code{\link{potato_visual}}, in which 50 % of the
+#' potatoes are initially unranked. Through ten subsequent timesteps, more
+#' ranks are added each time, given complete rankings at time 11. The dataset
+#' is used to demonstrate \code{\link{smc_mallows_new_item_rank}} and
+#' \code{\link{smc_mallows_update}}.
+#'
+#' @family datasets
+#' @references \insertRef{liu2019}{BayesMallows}
+#'
+"potato_partial"