+++ R/smc_mallows_new_item_rank.R
@@ -3,7 +3,6 @@
 #'   a new item ranks from an existing user at each time step. Each correction
 #'   and augmentation is done by filling in the missing item ranks using
 #'   pseudolikelihood augmentation.
-#' @param n_items Integer is the number of items in a ranking
 #' @param rankings 3D array of size n_assessors by n_items by timesteps containing a
 #'   set of observed rankings of timesteps time steps
 #' @param metric A character string specifying the distance metric to use in the
@@ -13,7 +12,6 @@
 #' @param leap_size leap_size Integer specifying the step size of the
 #'   leap-and-shift proposal distribution
 #' @param n_particles Integer specifying the number of particles
-#' @param timesteps Integer specifying the number of time steps in the SMC algorithm
 #' @param logz_estimate Estimate of the partition function, computed with
 #'   \code{\link{estimate_partition_function}}. Be aware that when using an
 #'   estimated partition function when \code{n_clusters > 1}, the partition
@@ -59,10 +57,8 @@
 #'
 #'
 smc_mallows_new_item_rank <- function(
-    n_items,
     rankings,
     n_particles,
-    timesteps,
     logz_estimate = NULL,
     mcmc_kernel_app,
     aug_rankings_init = NULL,
@@ -81,6 +77,9 @@ smc_mallows_new_item_rank <- function(
     "footrule", "spearman", "cayley", "hamming",
     "kendall", "ulam"
   ))
+
+  n_items <- dim(rankings)[[2]]
+  timesteps <- dim(rankings)[[3]]
 
   logz_list <- prepare_partition_function(logz_estimate, metric, n_items)