+++ tests/testthat/test-smc_mallows_new_item_rank.R
@@ -23,8 +23,8 @@ alpha_max <- 1e6
 test_that("Produces the wrong metric and aug_method error", {
   expect_error(
     smc_mallows_new_item_rank(
-      alpha = alpha_0, n_items = n_items, rankings = sample_dataset,
-      metric = "cayley", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      alpha = alpha_0, rankings = sample_dataset,
+      metric = "cayley", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "pseudolikelihood",
@@ -34,8 +34,8 @@ test_that("Produces the wrong metric and aug_method error", {
   )
   expect_error(
     smc_mallows_new_item_rank(
-      n_items = n_items, rankings = sample_dataset,
-      metric = "cayley", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      rankings = sample_dataset,
+      metric = "cayley", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "pseudolikelihood"
@@ -48,8 +48,8 @@ test_that("Runs with unif kernel", {
   set.seed(0)
   smc_unif_alpha_fixed_unif <- suppressMessages(
     smc_mallows_new_item_rank(
-      alpha = alpha_0, n_items = n_items, rankings = sample_dataset,
-      metric = "footrule", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      alpha = alpha_0, rankings = sample_dataset,
+      metric = "footrule", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "random", alpha_fixed = TRUE
@@ -65,8 +65,8 @@ test_that("Runs with unif kernel", {
   set.seed(2)
   smc_unif <- suppressMessages(
     smc_mallows_new_item_rank(
-      n_items = n_items, rankings = sample_dataset,
-      metric = "footrule", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      rankings = sample_dataset,
+      metric = "footrule", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "random"
@@ -81,8 +81,8 @@ test_that("Runs with unif kernel", {
 test_that("Runs with pseudo kernel", {
   smc_unif_alpha_fixed_unif <- suppressMessages(
     smc_mallows_new_item_rank(
-      alpha = alpha_0, n_items = n_items, rankings = sample_dataset,
-      metric = "footrule", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      alpha = alpha_0, rankings = sample_dataset,
+      metric = "footrule", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "pseudolikelihood",
@@ -96,8 +96,8 @@ test_that("Runs with pseudo kernel", {
   set.seed(1)
   smc_unif <- suppressMessages(
     smc_mallows_new_item_rank(
-      n_items = n_items, rankings = sample_dataset,
-      metric = "footrule", leap_size = leap_size, n_particles = n_particles, timesteps = timesteps,
+      rankings = sample_dataset,
+      metric = "footrule", leap_size = leap_size, n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app,
       alpha_prop_sd = alpha_prop_sd, lambda = lambda,
       alpha_max = alpha_max, aug_method = "pseudolikelihood"