+++ tests/testthat/test-smc_updated_new_item_rank.R
@@ -50,9 +50,9 @@ smc_test_new_user_unif <- smc_mallows_new_users(
 
 # run smc updated rankings with alpha unknown
 smc_test_partial_unif1 <- smc_mallows_new_item_rank(
-  alpha = 2, n_items = n_items,
+  alpha = 2,
   rankings = test_dataset, metric = metric, leap_size = leap_size,
-  n_particles = n_particles, timesteps = timesteps2,
+  n_particles = n_particles,
   mcmc_kernel_app = mcmc_kernel_app, aug_method = "random",
   rho_samples_init = smc_test_new_user_unif$rho_samples[, , timesteps + 1],
   aug_rankings_init = smc_test_new_user_unif$augmented_rankings,
@@ -68,9 +68,8 @@ test_that("Updated item rank output is OK", {
 
 # run smc updated rankings with alpha unknown
 smc_test_partial_unif2 <- smc_mallows_new_item_rank(
-  n_items = n_items,
   rankings = test_dataset, metric = metric, leap_size = leap_size,
-  n_particles = n_particles, timesteps = timesteps2,
+  n_particles = n_particles,
   mcmc_kernel_app = mcmc_kernel_app, alpha_prop_sd = 0.5,
   lambda = 0.1, alpha_max = 20, aug_method = "random",
   alpha_samples_init = smc_test_new_user_unif$alpha_samples[, timesteps + 1],
@@ -98,9 +97,9 @@ smc_test_new_user_pseudo <- smc_mallows_new_users(
 )
 
 smc_test_partial_pseudo1 <- smc_mallows_new_item_rank(
-  alpha = 2, n_items = n_items,
+  alpha = 2,
   rankings = test_dataset, metric = metric, leap_size = leap_size,
-  n_particles = n_particles, timesteps = timesteps2,
+  n_particles = n_particles,
   mcmc_kernel_app = mcmc_kernel_app, aug_method = "pseudolikelihood",
   rho_samples_init = smc_test_new_user_pseudo$rho_samples[, , timesteps + 1],
   aug_rankings_init = smc_test_new_user_pseudo$augmented_rankings,
@@ -115,9 +114,8 @@ test_that("Updated item rank output is OK", {
 })
 
 smc_test_partial_pseudo2 <- smc_mallows_new_item_rank(
-  n_items = n_items,
   rankings = test_dataset, metric = metric, leap_size = leap_size,
-  n_particles = n_particles, timesteps = timesteps2,
+  n_particles = n_particles,
   mcmc_kernel_app = mcmc_kernel_app, alpha_prop_sd = 0.5,
   lambda = 0.1, alpha_max = 20, aug_method = "pseudolikelihood",
   alpha_samples_init = smc_test_new_user_unif$alpha_samples[, timesteps + 1],
@@ -137,9 +135,9 @@ test_that("Updated item rank output (variable alpha) is OK", {
 test_that("metric and aug_method must match", {
   expect_error(
     smc_mallows_new_item_rank(
-      alpha = 2, n_items = n_items,
+      alpha = 2,
       rankings = test_dataset, metric = "cayley", leap_size = leap_size,
-      n_particles = n_particles, timesteps = timesteps2,
+      n_particles = n_particles,
       mcmc_kernel_app = mcmc_kernel_app, aug_method = "pseudolikelihood",
       alpha_fixed = TRUE
     ),